Cleanups for language and functionality
Shrink Screen and Enlarge Screen are now Shrink Hud and Enlarge Hud and appropriate controls are assigned to each (- will reduce the hud size and increase your view and = or + will make the hud bigger and decrease your view size).

Fixes for Incinerator and Freeze Ray.  They now act like the Expander and the Shrink Ray instead of the Incinerator being ignored in the status bar.

Typo fixed (DEVISTATOR vs DEVASTATOR)..  come on, English is easy.